requirements.txt
couchbase==2.0
wheel==0.23.0
python-dateutil==2.1
jsonpickle==0.4.0